fix(frontend): move isChunkLoadError to its own file to satisfy react-refresh/only-export-components
